Updated TODOs and removed the task of adding network ID to modules.
This is because it's hard to keep the ID updated e.g. when there are deletions that cause networks to split. If modules are not to be operated on for more than once when walk_network() is called multiple times (such as by SelectNetwork), then a hashmap of the object names should solve the problem.